@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-
-using AutoMapper;
 
 using Domain.Module;
 
@@ -19,70 +16,66 @@ using Infrastructure.Configurations;
 
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 using Trackers.ReadStore;
 using Trackers.ReadStore.Module;
 
-namespace Trackers.Service {
-    public class Startup {
-        private readonly IConfiguration _configuration;
+namespace Trackers.Service; 
 
-        public Startup(IConfiguration configuration) {
-            _configuration = configuration;
+public class Startup {
+    private readonly IConfiguration _configuration;
+
+    public Startup(IConfiguration configuration) {
+        _configuration = configuration;
+    }
+
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public IServiceProvider ConfigureServices(IServiceCollection services) {
+
+        var env = EnvironmentConfiguration.Bind(_configuration);
+
+        services
+            .AddSingleton(env)
+            .AddSwaggerGen(c => c.SwaggerDoc("v1", new OpenApiInfo { Title = "Trackers API", Version = "v1" }))
+            .AddMvc(e => e.EnableEndpointRouting = false);
+
+        return EventFlowOptions.New
+            .UseServiceCollection(services)
+            .AddAspNetCore()
+            .UseConsoleLog()
+            .RegisterModule<DomainModule>()
+            .RegisterModule<TrackingReadStoreModule>()
+            .RegisterModule<EventSourcingModule>()
+            .PublishToRabbitMq(RabbitMqConfiguration.With(new Uri(env.RabbitMqConnection)))
+            .CreateServiceProvider();
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env) {
+
+        // initialize InfoDbContext
+        using (var scope = app.ApplicationServices.CreateScope())
+        {
+            var dbContext = scope.ServiceProvider.GetService<IDbContextProvider<TrackingContext>>();
+            dbContext?.CreateContext();
         }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public IServiceProvider ConfigureServices(IServiceCollection services) {
-
-            var env = EnvironmentConfiguration.Bind(_configuration);
-
-            services
-                .AddAutoMapper()
-                .AddSingleton(env)
-                .AddSwaggerGen(c => c.SwaggerDoc("v1", new Info {Title = "Trackers API", Version = "v1"}))
-                .AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
-            return EventFlowOptions.New
-                .UseServiceCollection(services)
-                .AddAspNetCoreMetadataProviders()
-                .UseConsoleLog()
-                .RegisterModule<DomainModule>()
-                .RegisterModule<TrackingReadStoreModule>()
-                .RegisterModule<EventSourcingModule>()
-                .PublishToRabbitMq(RabbitMqConfiguration.With(new Uri(env.RabbitMqConnection)))
-                .CreateServiceProvider();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env) {
-
-            // initialize InfoDbContext
-            using (var scope = app.ApplicationServices.CreateScope())
-            {
-                var dbContext = scope.ServiceProvider.GetService<IDbContextProvider<TrackingContext>>();
-                dbContext.CreateContext();
-            }
-        
-
-            if (env.IsDevelopment()) {
-                app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-                app.UseSwaggerUI(c => { c.SwaggerEndpoint("/swagger/v1/swagger.json", "Trackers API V1"); });
-                app.UseMvc();
-            }
-            else {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-
-            app.UseHttpsRedirection();
+        if (env.IsDevelopment()) {
+            app.UseDeveloperExceptionPage();
+            app.UseSwagger();
+            app.UseSwaggerUI(c => { c.SwaggerEndpoint("/swagger/v1/swagger.json", "Trackers API V1"); });
             app.UseMvc();
         }
+        else {
+            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
+        }
+
+        app.UseHttpsRedirection();
+        app.UseMvc();
     }
 }
