@@ -5,19 +5,26 @@ using EventFlow.EntityFramework;
 using Infrastructure.Configurations;
 
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 
 namespace EventStore.Middleware {
     public class EventSourcingDbContextProvider : IDbContextProvider<EventSourcingDbContext>, IDisposable {
-        private readonly DbContextOptions<EventSourcingDbContext> _options;
 
-        public EventSourcingDbContextProvider(EnvironmentConfiguration configuration) {
-            _options = new DbContextOptionsBuilder<EventSourcingDbContext>()
-                .UseSqlServer(configuration.EventSourcingConnection)
-                .Options;
+        private readonly IConfiguration _configuration;
+
+        private DbContextOptions<EventSourcingDbContext> _options;
+
+        public EventSourcingDbContextProvider(IConfiguration configuration) {
+            _configuration = configuration;
         }
 
 
         public EventSourcingDbContext CreateContext() {
+            if(_options == null)
+                _options = new DbContextOptionsBuilder<EventSourcingDbContext>()
+                    .UseSqlServer(EnvironmentConfiguration.Bind(_configuration).EventSourcingConnection)
+                    .Options;
+            
             var db = new EventSourcingDbContext(_options);
             db.Database.EnsureCreated();
             return db;
